requirements.txt
# Django and core dependencies
Django==5.0.8
psycopg[binary]==3.2.1
django-environ==0.11.2
gunicorn==22.0.0
whitenoise==6.7.0

# Authentication
django-allauth==0.63.6

# State machine for orders
django-fsm==2.8.1

# Django REST Framework
djangorestframework==3.15.2
django-filter==24.2

# Development tools
django-debug-toolbar==4.4.6

# Testing
pytest==8.3.2
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.0

# Code quality
ruff==0.5.5
black==24.8.0
isort==5.13.2
mypy==1.11.1
django-stubs==5.0.4
types-requests==2.32.0.20240712

# Pre-commit hooks
pre-commit==3.8.0

# Utilities
python-decouple==3.8
Pillow==10.4.0